pyproject.toml
[tool.poetry]
name = "new_project"
version = "0.1.0"
description = ""
authors = ["Andrey Smakotin <slamfromm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^6.0.0"}
Pillow = "^9.0.0"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
